Import functions from a file, does some maths and prints the result

+++ 0x02-python-import_modules/calculator_1.py
@@ -48,4 +48,4 @@ def div(a, b):
     Returns:
            The return value. a / b
     """
-    return (a / b)
+    return int(a / b)
